requirements.txt
aiohttp==3.9.3
celery==5.3.6
Django==4.2.3
django-admin-thumbnails==0.2.8
django-cors-headers==4.3.1
django-filter==24.1
django-jazzmin==2.6.0
django-redis==5.4.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
drf-yasg==1.21.7
gunicorn==21.2.0
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.1
redis==5.0.3
requests==2.31.0
requests-oauthlib==1.4.0
